ref: simplify gaia nextjs config.
