+++ apps/gaia/tsconfig.json
@@ -20,6 +20,6 @@
     "next-env.d.ts",
     "src/**/*.ts",
     "src/**/*.tsx",
-    "tailwind.config.js"
+    "*.js"
   ]
 }
