+++ apps/gaia/version.js
@@ -1,0 +1,5 @@
+const fs = require("fs")
+
+const version = fs.readFileSync("../../VERSION", "UTF-8")
+
+module.exports = version
